requirements.txt
pandas
numpy
scikit-learn
python-dotenv
requests

streamlit
PyPDF2
python-docx

faiss-cpu
openai==0.28
sentence-transformers


